backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
supabase==2.9.1
python-dotenv==1.0.1
Pillow==10.4.0
pdf2image==1.17.0
groq==0.11.0
pydantic-ai==0.0.19
